3 Views added
MainView, LikedView & AboutView.

Now the upcoming - is to create the views.

WKWebkit for the about view.
LikedView needs a TableView with information from the Keyboard to the DataTable so you can see all your liked icebreakers

and then the home TAB where you get guided into how to activate the keyboard